refactor(library-3d): split 1392-line component into focused service/utils
Extract Three.js scene management into Library3dSceneService (436 lines)
and book mesh building into library-3d-mesh.utils.ts (221 lines). Move
CSS to a separate stylesheet. Component is now 359 lines (down from 1392),
delegating scene setup, animation, and mesh creation to extracted modules.

All public component inputs/outputs preserved unchanged.
